Wrap fetcher data access in Txn.calculateRead()
Due to how GraphQL Java dispatches GraphQL queries asynchronously some
of the data fetchers may run on a different thread than the one that
originated the query.  This is problematic for Jena datasets where
transactions are pinned to threads, therefore wrap all fetch logic that
accesses a DatasetGraph in Txn.calculateRead().  This is effectively a
no-op if running on a thread that already has an active transaction, and
creates a new one if necessary.  This fixes the intermittently seen "Not
in a transaction" error.