Renamed rangeFilter to nodeFilter (CORE-801)
This better reflects intent and purpose, of filtering by the nodes at
the other end of the relationships as oppposed to range which implied a
more specific kind of filtering.

Also removed domainFilter as that didn't make any practical sense as the
domain of inRels/outRels is always the subject we're traversing to/from
so there wasn't any value in this filter.